Changed folder strcuture
